fix: batch E polish — capabilities RPC, graceful shutdown, Clone error, u64 usage
- E1: ratd graceful shutdown via serve_with_shutdown (SIGTERM/SIGINT)
- E2: GetCapabilities RPC; ServiceClient queries server instead of hardcoding
- E3: ToolDefinition conversion warns on malformed JSON
- E4: fix secrets path in AGENTS.md (~/.ratatoskr/, not ~/.config/)
- E5: generate_stream uses filter_map, no spurious empty-text events
- E6: document retry_after as forward-looking on RateLimited
- E7: add label docs to PARAMETER_DISCOVERIES_TOTAL
- E8: save_cache tmp file includes PID to avoid collision
- E9: fetch_remote uses 30s timeout via reqwest::Client
- E10: seed.json uses versioned format; loader via parse_payload()
- E11: RatatoskrError is Clone (Json variant stores String)
- E12: document from_scores tie-breaking order
- E13: Usage fields u32 → u64 with proto conversion casts
- E14: cross-reference docs between set_parameters() impls
- E15: remove stale #[allow(dead_code)], document intentional ones
- E16: TODO → issue #21
- E17: remove duplicate test_capabilities_chat_only
- E18: update-deps creates chore branch instead of direct push
- E19: bump ort to 2.0.0-rc.11, document vergen-gitcl beta
- E20: add serde roundtrip tests for Usage, ChatEvent, ChatResponse

+++ src/convert/mod.rs
@@ -99,9 +99,9 @@ pub fn from_llm_tool_calls(calls: &[llm::ToolCall]) -> Vec<ToolCall> {
 /// Convert llm crate usage to our format
 pub fn from_llm_usage(usage: &llm::chat::Usage) -> RataUsage {
     RataUsage {
-        prompt_tokens: usage.prompt_tokens,
-        completion_tokens: usage.completion_tokens,
-        total_tokens: usage.prompt_tokens + usage.completion_tokens,
+        prompt_tokens: usage.prompt_tokens as u64,
+        completion_tokens: usage.completion_tokens as u64,
+        total_tokens: (usage.prompt_tokens + usage.completion_tokens) as u64,
         reasoning_tokens: None, // llm crate doesn't expose this yet
     }
 }
